// src/enums/resources.ts
export enum Resources {
  LISTING = "listing",
  AUTHORS = "authors",
  CATEGORIES = "categories",
  COMPANY_SUBMISSION = "company-submissions",
  LEAD_FORM_SUBMISSION = "lead-form-submissions",
  PAGES = "pages",
  POSTS = "posts",
  PRODUCTS = "products",
  PROJECTS = "projects",
  PRODUCT_CATEGORIES = "product-categories",
  PRODUCT_TAGS = "product-tags",
  PROPERTY = "properties",
  PROPERTY_LOCATIONS = "property-locations",
  PROPERTY_TYPES = "property-types",
  SEARCHES = "searches",
  SECTIONS = "sections",
  SERVICES = "services",
  FEATURES = "features",
  TAGS = "tags",
  USERS = "users",
}